[#780] Move error enum generator to common
